Checklist update - mostly done
Confirming which checkmarks we still have. Most that are left unchecked need just a bit more beef, a second glance, and I have one more test that i'll write at least in addition to the hard coded addresses vs csv addresses, need to confirm that we have both accounted for.